refactor contract parsing: dataclasses, split into modules
